requirements.txt
pre-commit==3.2.2
black==23.3.0
flake8==6.0.0
pytype==2022.4.26
timm==0.6.13
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
absl-py==1.4.0
lightning~=2.0.1
torchmetrics~=0.11.4